Added option to run tests with a different JVM than the build one.
