--- apm-agent-java-17/apm-agent-spring-6/pom.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>co.elastic.apm</groupId>
-        <artifactId>apm-agent-java-17</artifactId>
-        <version>1.36.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>apm-agent-spring-6</artifactId>
-    <name>${project.groupId}:${project.artifactId}</name>
-    <properties>
-        <!-- for licence header plugin -->
-        <apm-agent-parent.base.dir>${project.basedir}/../..</apm-agent-parent.base.dir>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>apm-agent-core</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-            <version>6.0.1</version>
-            <scope>provided</scope>
-        </dependency>
-        <!-- required for the creation of a mock response -->
-        <dependency>
-            <groupId>io.projectreactor</groupId>
-            <artifactId>reactor-core</artifactId>
-            <version>3.2.0.RELEASE</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-</project>
