--- apm-agent-java-17/pom.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <artifactId>apm-agent-parent</artifactId>
-        <groupId>co.elastic.apm</groupId>
-        <version>1.36.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>apm-agent-java-17</artifactId>
-    <packaging>pom</packaging>
-    <name>${project.groupId}:${project.artifactId}</name>
-
-    <modules>
-        <module>apm-agent-spring-6</module>
-    </modules>
-
-    <properties>
-        <maven-deploy-plugin.skip>true</maven-deploy-plugin.skip>
-        <!-- for licence header plugin -->
-        <apm-agent-parent.base.dir>${project.basedir}/..</apm-agent-parent.base.dir>
-
-        <!-- fails on inspection of Java 17 bytecode -->
-        <animal.sniffer.skip>true</animal.sniffer.skip>
-    </properties>
-
-</project>
