+++ apm-agent/pom.xml
@@ -420,22 +420,6 @@
         </dependency>
     </dependencies>
 
-    <profiles>
-        <profile>
-            <id>non-java-11</id>
-            <activation>
-                <jdk>!11</jdk>
-            </activation>
-            <dependencies>
-                <dependency>
-                    <groupId>${project.groupId}</groupId>
-                    <artifactId>apm-agent-spring-6</artifactId>
-                    <version>${project.version}</version>
-                </dependency>
-            </dependencies>
-        </profile>
-    </profiles>
-
     <build>
         <plugins>
             <plugin>
