+++ pom.xml
@@ -220,32 +220,45 @@
           </properties>
           <build>
             <plugins>
-              <plugin>
-                  <artifactId>maven-failsafe-plugin</artifactId>
-                  <configuration>
-                      <skipTests>${skip.integration.test}</skipTests>
-                      <rerunFailingTestsCount>3</rerunFailingTestsCount>
-                  </configuration>
-              </plugin>
-              <plugin>
-                  <artifactId>maven-surefire-plugin</artifactId>
-                  <configuration>
-                      <skipTests>true</skipTests>
-                  </configuration>
-              </plugin>
+                <plugin>
+                    <artifactId>maven-failsafe-plugin</artifactId>
+                    <configuration>
+                        <skipTests>${skip.integration.test}</skipTests>
+                        <rerunFailingTestsCount>3</rerunFailingTestsCount>
+                    </configuration>
+                </plugin>
+                <plugin>
+                    <artifactId>maven-surefire-plugin</artifactId>
+                    <configuration>
+                        <skipTests>true</skipTests>
+                    </configuration>
+                </plugin>
             </plugins>
           </build>
         </profile>
         <profile>
-          <!-- this profile should actually be active whenever running on Java 17 or higher, but since the only pre-Java-17 version we
-          test with is 11, this should be sufficient -->
-          <id>non-java-11</id>
-          <activation>
-              <jdk>!11</jdk>
-          </activation>
-          <modules>
-              <module>apm-agent-java-17</module>
-          </modules>
+            <id>different-test-jdk</id>
+            <activation>
+                <property>
+                    <name>test_java_binary</name>
+                </property>
+            </activation>
+            <build>
+                <plugins>
+                    <plugin>
+                        <artifactId>maven-surefire-plugin</artifactId>
+                        <configuration>
+                            <jvm>${test_java_binary}</jvm>
+                        </configuration>
+                    </plugin>
+                    <plugin>
+                        <artifactId>maven-failsafe-plugin</artifactId>
+                        <configuration>
+                            <jvm>${test_java_binary}</jvm>
+                        </configuration>
+                    </plugin>
+                </plugins>
+            </build>
         </profile>
         <profile>
             <id>no-errorprone</id>
@@ -433,7 +446,6 @@
                     </execution>
                 </executions>
             </plugin>
-
             <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
